modulo geometry para as classes Point e Directions
