+++ snake/characters.py
@@ -1,25 +1,10 @@
 from collections import deque, namedtuple
 
-
-class Point(namedtuple('Point', ['x', 'y'], defaults=[0, 0])):
-    __slots__ = ()
-
-    def __add__(self, other):
-        return Point(self.x + other.x, self.y + other.y)
-
-    def __mul__(self, scalar):
-        return Point(self.x * scalar, self.y * scalar)
+from snake.geometry import Point, Directions
 
 
 class Apple(Point):
     pass
-
-
-class Directions:
-    UP = Point(0, -1)
-    DOWN = Point(0, 1)
-    RIGHT = Point(1, 0)
-    LEFT = Point(-1, 0)
 
 
 class Snake:
